Allow the Oracle to figure out which plasma blocks to send to mainnet
The Oracle can look at the current plasma block numbers in the Solidity
and Go contracts, and then use the plasma block numbering scheme to
iterate through the finalized but as yet unsubmitted to mainnet
non-deposit plasma blocks in the DAppChain.

This approach doesn't require trawling through DAppChain or Ethereum
logs.